Refactoring to just use regular beans
The NoSQLEntity was getting pretty complicated and made it more
difficult for a user to save their data. Now we can just use any regular
bean and have gson do the serialization and deserialization for you
without having to convert your data into a NoSQLEntity manually. This
also means you won't have to deserialize it into your beans afterwards.

The main concern now is that you'll need to make sure that any changes
to a bean are backwards compatible.