Added customizable Serializers/Deserializers
You can now use a custom serializer or deserializer with the NoSQL
library. By default it will still use Gson, but the user can supply
their own if they need to use Jackson or be more specific on what fields
should be stored or not.

Also, a license was added to the README. Javadocs were added as well.