Added more tests
There's now a test for a Null object being inserted into the datastore
which works fine. In addition, there's a test of an "old" data model
being stored and then restored into a "new" data model which works as
long as the new contains all of the fields that the old model had.