Grudging support for synchronous UI calls.
You should never get synchronous results of a data operation on the UI thread (it locks the UI while the data is being
completed). While some may say we should give the developers the option to do whatever they want, by default this
behavior is frankly stupid and won't be supported.

If you truly believe that you should be making synchronous calls on the UI thread, you may create your NoSQL object
now by supplying a QueryDelivery on a different Handler thread (this could be useful for services as well). An error
message will still be generated in the logs if you attempt this.