Store data as a byte[] rather than a String
To make this library more flexible, data is now stored as a byte[] rather than a string. This will let you store any
arbitrary piece of data, not just data that easily serializes as a String. This is an API breaking change, but data
should be retained.

The regular DB upgrade procedure will update the database to using a byte array. Users will just need to switch the API
of any custom serializers/deserializers to accept/return a byte[]. If you used the default GonSerializer, no changes
are required of you.